# core: cache Accept-Encoding headers