--- a/interop-testing/src/test/java/io/grpc/testing/integration/CompressionTest.java
+++ b/interop-testing/src/test/java/io/grpc/testing/integration/CompressionTest.java
@@ -66,6 +66,7 @@ import io.grpc.testing.integration.Messages.SimpleRequest;
 import io.grpc.testing.integration.Messages.SimpleResponse;
 import io.grpc.testing.integration.TestServiceGrpc.TestServiceBlockingStub;
 import io.grpc.testing.integration.TransportCompressionTest.Fzip;
+import java.nio.charset.Charset;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -231,13 +232,13 @@ public class CompressionTest {
     }
 
     if (serverAcceptEncoding) {
-      assertEquals("fzip", clientResponseHeaders.get(MESSAGE_ACCEPT_ENCODING_KEY));
+      assertEqualsString("fzip", clientResponseHeaders.get(MESSAGE_ACCEPT_ENCODING_KEY));
     } else {
       assertNull(clientResponseHeaders.get(MESSAGE_ACCEPT_ENCODING_KEY));
     }
 
     if (clientAcceptEncoding) {
-      assertEquals("fzip", serverResponseHeaders.get(MESSAGE_ACCEPT_ENCODING_KEY));
+      assertEqualsString("fzip", serverResponseHeaders.get(MESSAGE_ACCEPT_ENCODING_KEY));
     } else {
       assertNull(serverResponseHeaders.get(MESSAGE_ACCEPT_ENCODING_KEY));
     }
@@ -322,5 +323,9 @@ public class CompressionTest {
       headersCopy.merge(headers);
       clientResponseHeaders = headersCopy;
     }
+  }
+
+  private static void assertEqualsString(String expected, byte[] actual) {
+    assertEquals(expected, new String(actual, Charset.forName("US-ASCII")));
   }
 }